refactor(chessboard): split into components, dedupe colors, bundle avatars
Behaviour-preserving cleanup of the chessboard demo (index.tsx was ~800
lines, checkmate-aura.tsx ~815):

- Extract PlayerCard (+ CaptureTray), HistorySync, the post-game
  GameReviewCard, the wave SKSL shader, shared types, constants and
  utils into their own modules, matching the per-component structure
  used by the other demos.
- Add a withAlpha() theme helper and route the hand-written rgba
  literals (last-move highlight, result tags, history pill) through it
  so tinted fills stay sourced from the palette.
- Collapse the duplicate Color/Side type into a single Side.
- Bundle the player avatars locally instead of fetching pravatar URLs
  at runtime, so the demo renders fully offline.
- Guard the async replay / checkmate-aura timers against unmount.

Validated on an iOS dev client: replay, history scrubbing, checkmate
aura + review card, flip and avatars all behave as before.
